+++ src/main/java/net/ccbluex/liquidbounce/injection/forge/mixins/gui/MixinGuiButtonExt.java
@@ -76,7 +76,7 @@ public abstract class MixinGuiButtonExt extends GuiButton {
             moveX = AnimationUtils.animate(0F, moveX, speedDelta);
          }
 
-         float roundCorner = (float) Math.max(0F, 2.4F + moveX - (this.width - 2.4F));
+         float roundCorner = Math.max(0F, 2.4F + moveX - (this.width - 2.4F));
 
          switch (hud.getGuiButtonStyle().get().toLowerCase()) {
             case "minecraft":
