+++ src/main/java/net/ccbluex/liquidbounce/utils/Colors.java
@@ -75,7 +75,7 @@ public class Colors
     }
 
     public static Color astolfoRainbow(int delay,int offset, int index) {
-        double rainbowDelay = Math.ceil(System.currentTimeMillis() + (long)(delay * index)) / offset;
+        double rainbowDelay = Math.ceil(System.currentTimeMillis() + (long)((long) delay * index)) / offset;
         return Color.getHSBColor((double)((float)((rainbowDelay %= 360.0) / 360.0)) < 0.5 ? -((float)(rainbowDelay / 360.0)) : (float)(rainbowDelay / 360.0), 0.5F, 1.0F);
     }
